Revert unrelated commits
