+++ modules/router-store/src/index.ts
@@ -24,14 +24,10 @@ export {
   NavigationActionTiming,
   ROUTER_CONFIG,
   DEFAULT_ROUTER_FEATURENAME,
-  RouterState,
 } from './router_store_module';
 export {
   RouterStateSerializer,
   DefaultRouterStateSerializer,
   SerializedRouterStateSnapshot,
   BaseRouterStoreState,
-  MinimalActivatedRouteSnapshot,
-  MinimalRouterStateSnapshot,
-  MinimalRouterStateSerializer,
-} from './serializers';
+} from './serializer';
