+++ modules/router-store/src/router_store_module.ts
@@ -31,21 +31,11 @@ import {
   RouterStateSerializer,
   SerializedRouterStateSnapshot,
   BaseRouterStoreState,
-  MinimalRouterStateSerializer,
-} from './serializers';
+} from './serializer';
 
 export type StateKeyOrSelector<
   T extends BaseRouterStoreState = SerializedRouterStateSnapshot
 > = string | Selector<any, RouterReducerState<T>>;
-
-/**
- * Full = Serializes the router event with DefaultRouterStateSerializer
- * Minimal = Serializes the router event with MinimalRouterStateSerializer
- */
-export const enum RouterState {
-  Full,
-  Minimal,
-}
 
 export interface StoreRouterConfig<
   T extends BaseRouterStoreState = SerializedRouterStateSnapshot
@@ -60,12 +50,6 @@ export interface StoreRouterConfig<
    * set this property to NavigationActionTiming.PostActivation.
    */
   navigationActionTiming?: NavigationActionTiming;
-  /**
-   * Decides which router serializer should be used, if there is none provided, and the metadata on the dispatched @ngrx/router-store action payload.
-   * Set to `Full` to use the `DefaultRouterStateSerializer` and to set the angular router events as payload.
-   * Set to `Minimal` to use the `MinimalRouterStateSerializer` and to set a minimal router event with the navigation id and url as payload.
-   */
-  routerState?: RouterState;
 }
 
 interface StoreRouterActionPayload {
@@ -166,9 +150,7 @@ export class StoreRouterConnectingModule {
           provide: RouterStateSerializer,
           useClass: config.serializer
             ? config.serializer
-            : config.routerState === RouterState.Minimal
-              ? MinimalRouterStateSerializer
-              : DefaultRouterStateSerializer,
+            : DefaultRouterStateSerializer,
         },
       ],
     };
@@ -325,10 +307,6 @@ export class StoreRouterConnectingModule {
         payload: {
           routerState: this.routerState,
           ...payload,
-          event:
-            this.config.routerState === RouterState.Minimal
-              ? { id: payload.event.id, url: payload.event.url }
-              : payload.event,
         },
       });
     } finally {
