+++ modules/router-store/src/serializer.ts
@@ -1,5 +1,19 @@
 import { ActivatedRouteSnapshot, RouterStateSnapshot } from '@angular/router';
-import { BaseRouterStoreState, RouterStateSerializer } from './shared';
+
+/**
+ * Simple router state.
+ * All custom router states / state serializers should have at least
+ * the properties of this interface.
+ */
+export interface BaseRouterStoreState {
+  url: string;
+}
+
+export abstract class RouterStateSerializer<
+  T extends BaseRouterStoreState = BaseRouterStoreState
+> {
+  abstract serialize(routerState: RouterStateSnapshot): T;
+}
 
 export interface SerializedRouterStateSnapshot extends BaseRouterStoreState {
   root: ActivatedRouteSnapshot;
