--- modules/router-store/src/serializers/index.ts
@@ -1,12 +1,0 @@
-export { BaseRouterStoreState, RouterStateSerializer } from './shared';
-
-export {
-  SerializedRouterStateSnapshot,
-  DefaultRouterStateSerializer,
-} from './default_serializer';
-
-export {
-  MinimalActivatedRouteSnapshot,
-  MinimalRouterStateSnapshot,
-  MinimalRouterStateSerializer,
-} from './minimal_serializer';
