+++ modules/store/spec/BUILD
@@ -4,13 +4,12 @@ ts_test_library(
     name = "test_lib",
     srcs = glob(
         [
-            "*.ts",
-            "meta-reducers/**/*.ts",
+            "**/*.ts",
         ],
+        exclude = ["ngc/**/*.ts"],
     ),
     deps = [
         "//modules/store",
-        "//modules/store/spec/fixtures",
         "//modules/store/testing",
         "@npm//rxjs",
         "@npm//ts-snippet",
