+++ modules/store/spec/store.spec.ts
@@ -9,6 +9,7 @@ import {
   select,
   ReducerManagerDispatcher,
   UPDATE,
+  REDUCER_FACTORY,
   ActionReducer,
   Action,
 } from '../';
@@ -22,7 +23,10 @@ import {
   counterReducer2,
 } from './fixtures/counter';
 import Spy = jasmine.Spy;
-import { take } from 'rxjs/operators';
+import any = jasmine.any;
+import { skip, take } from 'rxjs/operators';
+import { MockStore, provideMockStore } from '../testing';
+import { createSelector } from '../src/selector';
 
 interface TestAppSchema {
   counter1: number;
@@ -441,6 +445,180 @@ describe('ngRx Store', () => {
         reducerFactory: jasmine.createSpy(`reducerFactory_${key}`),
       };
     }
+  });
+
+  describe('Mock Store', () => {
+    let mockStore: MockStore<TestAppSchema>;
+    const initialState = { counter1: 0, counter2: 1, counter4: 3 };
+    const stringSelector = 'counter4';
+    const memoizedSelector = createSelector(
+      () => initialState,
+      state => state.counter4
+    );
+
+    beforeEach(() => {
+      TestBed.configureTestingModule({
+        providers: [
+          provideMockStore({
+            initialState,
+            selectors: [
+              { selector: stringSelector, value: 87 },
+              { selector: memoizedSelector, value: 98 },
+            ],
+          }),
+        ],
+      });
+
+      mockStore = TestBed.get(Store);
+    });
+
+    it('should set the initial state to a mocked one', (done: DoneFn) => {
+      const fixedState = {
+        counter1: 17,
+        counter2: 11,
+        counter3: 25,
+      };
+      mockStore.setState(fixedState);
+      mockStore.pipe(take(1)).subscribe({
+        next(val) {
+          expect(val).toEqual(fixedState);
+        },
+        error: done.fail,
+        complete: done,
+      });
+    });
+
+    it('should allow tracing dispatched actions', () => {
+      const action = { type: INCREMENT };
+      mockStore.scannedActions$
+        .pipe(skip(1))
+        .subscribe(scannedAction => expect(scannedAction).toEqual(action));
+      mockStore.dispatch(action);
+    });
+
+    it('should allow mocking of store.select with string selector using provideMockStore', () => {
+      const expectedValue = 87;
+
+      mockStore
+        .select(stringSelector)
+        .subscribe(result => expect(result).toBe(expectedValue));
+    });
+
+    it('should allow mocking of store.select with a memoized selector using provideMockStore', () => {
+      const expectedValue = 98;
+
+      mockStore
+        .select(memoizedSelector)
+        .subscribe(result => expect(result).toBe(expectedValue));
+    });
+
+    it('should allow mocking of store.pipe(select()) with a memoized selector using provideMockStore', () => {
+      const expectedValue = 98;
+
+      mockStore
+        .pipe(select(memoizedSelector))
+        .subscribe(result => expect(result).toBe(expectedValue));
+    });
+
+    it('should allow mocking of store.select with string selector using overrideSelector', () => {
+      const mockValue = 5;
+
+      mockStore.overrideSelector('counter1', mockValue);
+
+      mockStore
+        .select('counter1')
+        .subscribe(result => expect(result).toBe(mockValue));
+    });
+
+    it('should allow mocking of store.select with a memoized selector using overrideSelector', () => {
+      const mockValue = 5;
+      const selector = createSelector(
+        () => initialState,
+        state => state.counter1
+      );
+
+      mockStore.overrideSelector(selector, mockValue);
+
+      mockStore
+        .select(selector)
+        .subscribe(result => expect(result).toBe(mockValue));
+    });
+
+    it('should allow mocking of store.pipe(select()) with a memoized selector using overrideSelector', () => {
+      const mockValue = 5;
+      const selector = createSelector(
+        () => initialState,
+        state => state.counter2
+      );
+
+      mockStore.overrideSelector(selector, mockValue);
+
+      mockStore
+        .pipe(select(selector))
+        .subscribe(result => expect(result).toBe(mockValue));
+    });
+
+    it('should pass through unmocked selectors', () => {
+      const mockValue = 5;
+      const selector = createSelector(
+        () => initialState,
+        state => state.counter1
+      );
+      const selector2 = createSelector(
+        () => initialState,
+        state => state.counter2
+      );
+      const selector3 = createSelector(
+        selector,
+        selector2,
+        (sel1, sel2) => sel1 + sel2
+      );
+
+      mockStore.overrideSelector(selector, mockValue);
+
+      mockStore
+        .pipe(select(selector2))
+        .subscribe(result => expect(result).toBe(1));
+      mockStore
+        .pipe(select(selector3))
+        .subscribe(result => expect(result).toBe(6));
+    });
+
+    it('should allow you reset mocked selectors', () => {
+      const mockValue = 5;
+      const selector = createSelector(
+        () => initialState,
+        state => state.counter1
+      );
+      const selector2 = createSelector(
+        () => initialState,
+        state => state.counter2
+      );
+      const selector3 = createSelector(
+        selector,
+        selector2,
+        (sel1, sel2) => sel1 + sel2
+      );
+
+      mockStore
+        .pipe(select(selector3))
+        .subscribe(result => expect(result).toBe(1));
+
+      mockStore.overrideSelector(selector, mockValue);
+      mockStore.overrideSelector(selector2, mockValue);
+      selector3.release();
+
+      mockStore
+        .pipe(select(selector3))
+        .subscribe(result => expect(result).toBe(10));
+
+      mockStore.resetSelectors();
+      selector3.release();
+
+      mockStore
+        .pipe(select(selector3))
+        .subscribe(result => expect(result).toBe(1));
+    });
   });
 
   describe('Meta Reducers', () => {
