--- modules/store/testing/spec/BUILD
@@ -1,22 +1,0 @@
-load("//tools:defaults.bzl", "jasmine_node_test", "ts_test_library")
-
-ts_test_library(
-    name = "test_lib",
-    srcs = glob(
-        ["**/*.ts"],
-    ),
-    deps = [
-        "//modules/store",
-        "//modules/store/spec/fixtures",
-        "//modules/store/testing",
-        "@npm//rxjs",
-    ],
-)
-
-jasmine_node_test(
-    name = "test",
-    deps = [
-        ":test_lib",
-        "//modules/store",
-    ],
-)
