+++ projects/example-app/src/app/app-routing.module.ts
@@ -1,8 +1,7 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
-
 import { AuthGuard } from '@example-app/auth/services';
-import { NotFoundPageComponent } from '@example-app/core/containers';
+import { NotFoundPageComponent } from '@example-app/core/containers/not-found-page.component';
 
 export const routes: Routes = [
   { path: '', redirectTo: '/books', pathMatch: 'full' },
@@ -11,11 +10,7 @@ export const routes: Routes = [
     loadChildren: '@example-app/books/books.module#BooksModule',
     canActivate: [AuthGuard],
   },
-  {
-    path: '**',
-    component: NotFoundPageComponent,
-    data: { title: 'Not found' },
-  },
+  { path: '**', component: NotFoundPageComponent },
 ];
 
 @NgModule({
