+++ projects/example-app/src/app/app.module.ts
@@ -6,17 +6,16 @@ import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 
 import { StoreModule } from '@ngrx/store';
 import { EffectsModule } from '@ngrx/effects';
-import { StoreRouterConnectingModule, RouterState } from '@ngrx/router-store';
+import { StoreRouterConnectingModule } from '@ngrx/router-store';
 import { StoreDevtoolsModule } from '@ngrx/store-devtools';
 
+import { CoreModule } from '@example-app/core/core.module';
 import { AuthModule } from '@example-app/auth';
 
 import { ROOT_REDUCERS, metaReducers } from '@example-app/reducers';
 
-import { CoreModule } from '@example-app/core';
+import { AppComponent } from '@example-app/core/containers/app.component';
 import { AppRoutingModule } from '@example-app/app-routing.module';
-import { UserEffects, RouterEffects } from '@example-app/core/effects';
-import { AppComponent } from '@example-app/core/containers';
 
 @NgModule({
   imports: [
@@ -38,17 +37,13 @@ import { AppComponent } from '@example-app/core/containers';
       metaReducers,
       runtimeChecks: {
         strictImmutability: true,
-        strictStateSerializability: true,
-        strictActionSerializability: true,
       },
     }),
 
     /**
      * @ngrx/router-store keeps router state up-to-date in the store.
      */
-    StoreRouterConnectingModule.forRoot({
-      routerState: RouterState.Minimal,
-    }),
+    StoreRouterConnectingModule.forRoot(),
 
     /**
      * Store devtools instrument the store retaining past versions of state
@@ -74,7 +69,8 @@ import { AppComponent } from '@example-app/core/containers';
      *
      * See: https://ngrx.io/guide/effects#registering-root-effects
      */
-    EffectsModule.forRoot([UserEffects, RouterEffects]),
+    EffectsModule.forRoot([]),
+
     CoreModule,
   ],
   bootstrap: [AppComponent],
