+++ projects/example-app/src/app/auth/effects/auth.effects.ts
@@ -12,7 +12,6 @@ import {
 import { Credentials } from '@example-app/auth/models';
 import { AuthService } from '@example-app/auth/services';
 import { LogoutConfirmationDialogComponent } from '@example-app/auth/components';
-import { UserActions } from '@example-app/core/actions';
 
 @Injectable()
 export class AuthEffects {
@@ -67,13 +66,6 @@ export class AuthEffects {
             ? AuthActions.logout()
             : AuthActions.logoutConfirmationDismiss()
       )
-    )
-  );
-
-  logoutIdleUser$ = createEffect(() =>
-    this.actions$.pipe(
-      ofType(UserActions.idleTimeout),
-      map(() => AuthActions.logout())
     )
   );
