+++ projects/example-app/src/app/books/books-routing.module.ts
@@ -9,22 +9,13 @@ import {
 import { BookExistsGuard } from '@example-app/books/guards';
 
 export const routes: Routes = [
-  {
-    path: 'find',
-    component: FindBookPageComponent,
-    data: { title: 'Find book' },
-  },
+  { path: 'find', component: FindBookPageComponent },
   {
     path: ':id',
     component: ViewBookPageComponent,
     canActivate: [BookExistsGuard],
-    data: { title: 'Book details' },
   },
-  {
-    path: '',
-    component: CollectionPageComponent,
-    data: { title: 'Collection' },
-  },
+  { path: '', component: CollectionPageComponent },
 ];
 
 @NgModule({
