+++ projects/example-app/src/app/core/actions/index.ts
@@ -1,4 +1,3 @@
 import * as LayoutActions from './layout.actions';
-import * as UserActions from './user.actions';
 
-export { LayoutActions, UserActions };
+export { LayoutActions };
