--- projects/example-app/src/app/core/actions/user.actions.ts
@@ -1,3 +1,0 @@
-import { createAction } from '@ngrx/store';
-
-export const idleTimeout = createAction('[User] Idle Timeout');
