--- projects/example-app/src/app/core/components/index.ts
@@ -1,4 +1,0 @@
-export * from './layout.component';
-export * from './nav-item.component';
-export * from './sidenav.component';
-export * from './toolbar.component';
