--- projects/example-app/src/app/core/containers/index.ts
@@ -1,2 +1,0 @@
-export * from './app.component';
-export * from './not-found-page.component';
