--- projects/example-app/src/app/core/effects/index.ts
@@ -1,2 +1,0 @@
-export * from './user.effects';
-export * from './router.effects';
