+++ projects/example-app/src/app/core/reducers/layout.reducer.ts
@@ -1,5 +1,4 @@
 import { createReducer, on } from '@ngrx/store';
-
 import { LayoutActions } from '@example-app/core/actions';
 
 export interface State {
