+++ projects/ngrx.io/content/examples/store/src/app/counter.reducer.ts
@@ -1,11 +1,21 @@
 // #docregion
-import { createReducer, on } from '@ngrx/store';
+import { Action } from '@ngrx/store';
 import { increment, decrement, reset } from './counter.actions';
 
 export const initialState = 0;
 
-export const counterReducer = createReducer(initialState,
-  on(increment, state => state + 1),
-  on(decrement, state => state - 1),
-  on(reset, state => 0),
-);
+export function counterReducer(state = initialState, action: Action) {
+  switch (action.type) {
+    case increment.type:
+      return state + 1;
+
+    case decrement.type:
+      return state - 1;
+
+    case reset.type:
+      return 0;
+
+    default:
+      return state;
+  }
+}
