+++ projects/ngrx.io/content/guide/entity/adapter.md
@@ -69,7 +69,13 @@ export const initialState: State = adapter.getInitialState({
   selectedUserId: null,
 });
 
-export const reducer = createReducer(initialState);
+export function reducer(state = initialState, action): State {
+  switch (action.type) {
+    default: {
+      return state;
+    }
+  }
+}
 </code-example>
 
 ## Adapter Collection Methods
@@ -100,7 +106,7 @@ export interface User {
 </code-example>
 
 <code-example header="user.actions.ts">
-import { createAction, props } from '@ngrx/store';
+import { createAction, props, union } from '@ngrx/store';
 import { Update } from '@ngrx/entity';
 
 import { User } from '../models/user.model';
@@ -118,6 +124,21 @@ export const deleteUsers = createAction('[User/API] Delete Users', props&lt;{ id
 export const deleteUsersByPredicate = createAction('[User/API] Delete Users By Predicate', props&lt;{ predicate: Predicate&lt;User&gt; }&gt;());
 export const clearUsers = createAction('[User/API] Clear Users');
 
+const all = union({
+  loadUsers,
+  addUser,
+  upsertUser,
+  addUsers,
+  upsertUsers,
+  updateUser,
+  updateUsers,
+  mapUsers,
+  deleteUser,
+  deleteUsers,
+  deleteUsersByPredicate,
+  clearUsers
+});
+export type Union = typeof all;
 </code-example>
 
 <code-example header="user.reducer.ts">
@@ -137,43 +158,61 @@ export const initialState: State = adapter.getInitialState({
   selectedUserId: null,
 });
 
-export const reducer = createReducer(initialState,
-  on(UserActions.addUser, state => {
-   return adapter.addOne(action.user, state)
-  }),
-  on(UserActions.upsertUser, state => {
-    return adapter.upsertOne(action.user, state);
-  }),
-  on(UserActions.addUsers, state => {
-    return adapter.addMany(action.users, state);
-  }),
-  on(UserActions.upsertUsers, state => {
-    return adapter.upsertUsers(action.users, state);
-  }),
-  on(UserActions.updateUser, state => {
-    return adapter.updateOne(action.user, state);
-  }),
-  on(UserActions.updateUsers, state => {
-    return adapter.updateMany(action.users, state);
-  }),
-  on(UserActions.mapUsers, state => {
-    return adapter.map(action.entityMap, state);
-  }),
-  on(UserActions.deleteUser, state => {
-    return adapter.removeOne(action.ids, state);
-  }),
-  on(UserActions.deleteUsers, state => {
-    return adapter.removeMany(action.ids, state);
-  }),
-  on(UserActions.deleteUsersByPredicate, state => {
-    return adapter.removeMany(action.predicate, state);
-  }),
-  on(UserActions.loadUsers, state => {
-    return adapter.addAll(action.users, state);
-  }),
-  on(UserActions.clearUsers, state => {
-    return adapter.removeAll({ ...state, selectedUserId: null });
-  }));
+export function reducer(state = initialState, action: UserActions.Union): State {
+  switch (action.type) {
+    case UserActions.addUser.type: {
+      return adapter.addOne(action.user, state);
+    }
+
+    case UserActions.upsertUser.type: {
+      return adapter.upsertOne(action.user, state);
+    }
+
+    case UserActions.addUsers.type: {
+      return adapter.addMany(action.users, state);
+    }
+
+    case UserActions.upsertUsers.type: {
+      return adapter.upsertMany(action.users, state);
+    }
+
+    case UserActions.updateUser.type: {
+      return adapter.updateOne(action.user, state);
+    }
+
+    case UserActions.updateUsers.type: {
+      return adapter.updateMany(action.users, state);
+    }
+
+    case UserActions.mapUsers.type: {
+      return adapter.map(action.entityMap, state);
+    }
+
+    case UserActions.deleteUser.type: {
+      return adapter.removeOne(action.id, state);
+    }
+
+    case UserActions.deleteUsers.type: {
+      return adapter.removeMany(action.ids, state);
+    }
+
+    case UserActions.deleteUsersByPredicate.type: {
+      return adapter.removeMany(action.predicate, state);
+    }
+
+    case UserActions.loadUsers.type: {
+      return adapter.addAll(action.users, state);
+    }
+
+    case UserActions.clearUsers.type: {
+      return adapter.removeAll({ ...state, selectedUserId: null });
+    }
+
+    default: {
+      return state;
+    }
+  }
+}
 
 export const getSelectedUserId = (state: State) => state.selectedUserId;
